+++ src/Database/ConnectionFactory.php
@@ -7,6 +7,17 @@ use Illuminate\Database\Connectors\ConnectionFactory as BaseConnectionFactory;
 
 class ConnectionFactory extends BaseConnectionFactory
 {
+    protected function getWriteConfig(array $config)
+    {
+        $writeConfig = parent::getWriteConfig($config);
+
+        if ($writeConfig['use_first_available_host'] ?? true) {
+            $writeConfig['host'] = $this->parseHosts($writeConfig)[0];
+        }
+
+        return $writeConfig;
+    }
+
     protected function parseHosts(array $config)
     {
         return (new HostParser())